fixed infos

+++ util/util/system_utils.h
@@ -42,6 +42,6 @@ AGROS_UTIL_API bool isProcessRunning(int pid);
 
 AGROS_UTIL_API QString cpuType();
 AGROS_UTIL_API int cpuNumberOfCores();
-AGROS_UTIL_API long getMemorySize();
+AGROS_UTIL_API size_t getMemorySize();
 
 #endif // SYSTEM_UTILS_H
